requirements.txt
slack-bolt==1.21.3
slack-sdk==3.33.5
anthropic==0.42.0
python-dotenv==1.0.1
aiohttp==3.11.11
requests==2.32.3
flask==3.1.0
